feature: Provide a default renderer for Html
This feature uses Maud, a nice crate to generate HTML. The only issue
is that Maud requires an unstable feature of the compiler, and
therefore has to be build with a nightly rustc.

By default, the feature is disabled and need to be enabled in the
Cargo.toml of the user.