refactor: Add the author of a reply to its template
If we consider the HTML use case, it makes sense not to propagate the
author of a dialogue or thought reply to the actual content of this
reply, thanks to how CSS works.

Other use case, including BBCode and LaTeX, makes it less easy to
style a reply based on its author.