feature: Support special typography when a character speaks twice
Consider the following ogmarkup snipput:

    [Bonjour,| dit-il.|](foo)

    [Je suis venu.](foo)

Both dialogue lines are from the character `foo'. Before this patch,
this would have been rendered as:

    « Bonjour, dit-il.

    — Je suis venu. »

This is clearly misleading.

At least in French, there is a special typographic approach to deal
with dialogue written as several paragraphs, and that is to use the
closing quotes in place of the long hyphen. With this patch, the
ogmarkup snippet is now rendered as:

    « Bonjour, dit-il.

    » Je suis venu. »

The English typography does not introduce any particular character, so
an equivalent text would be rendered as:

    “ Hi, he says.

    I came.”