fix: Support empty documents and documents with leading spaces
Before this patch, empty documents were parsed as a document with only
one empty ill-formed section. In addition, documents with leading
spaces also had their own issues.

Even though the solution implemented in this patch is not as clean as
one would have hoped, it improves the situation, and both classes of
documents are now parsed as users would expect.